updated student

+++ src/main/java/com/irms_service/repository/PersonRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.irms_service.entity.PersonEntity;
 
 @Repository
-public interface PersonRepository extends JpaRepository<PersonEntity, Integer>{
+public interface PersonRepository extends JpaRepository<PersonEntity, Long>{
 
 	
 }
