+++ src/main/java/com/irms_service/repository/StudentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.irms_service.entity.StudentEntity;
 
 @Repository
-public interface StudentRepository extends JpaRepository<StudentEntity, Integer>{
+public interface StudentRepository extends JpaRepository<StudentEntity, Long>{
 
 	
 }
